chore_: old router logic removed
