+++ services/wallet/router/router_send_type.go
@@ -1,24 +1,16 @@
 package router
 
 import (
-	"context"
-	"fmt"
 	"math/big"
 	"strings"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/status-im/status-go/eth-node/types"
 	"github.com/status-im/status-go/params"
-	"github.com/status-im/status-go/services/ens"
-	"github.com/status-im/status-go/services/stickers"
-	"github.com/status-im/status-go/services/wallet/bigint"
 	"github.com/status-im/status-go/services/wallet/collectibles"
 	walletCommon "github.com/status-im/status-go/services/wallet/common"
 	"github.com/status-im/status-go/services/wallet/market"
 	"github.com/status-im/status-go/services/wallet/router/pathprocessor"
 	"github.com/status-im/status-go/services/wallet/token"
-	"github.com/status-im/status-go/transactions"
 )
 
 type SendType int
@@ -89,18 +81,6 @@ func (s SendType) FindToken(tokenManager *token.Manager, collectibles *collectib
 		Decimals: 0,
 		ChainID:  network.ChainID,
 	}
-}
-
-// TODO: remove this function once we fully move to routerV2
-func (s SendType) isTransfer(routerV2Logic bool) bool {
-	return s == Transfer ||
-		s == Bridge && routerV2Logic ||
-		s == Swap ||
-		s.IsCollectiblesTransfer()
-}
-
-func (s SendType) needL1Fee() bool {
-	return !s.IsEnsTransfer() && !s.IsStickersTransfer()
 }
 
 // canUseProcessor is used to check if certain SendType can be used with a given path processor
@@ -194,45 +174,4 @@ func (s SendType) isAvailableFor(network *params.Network) bool {
 	}
 
 	return false
-}
-
-// TODO: remove this function once we fully move to routerV2
-func (s SendType) EstimateGas(ensService *ens.Service, stickersService *stickers.Service, network *params.Network, from common.Address, tokenID string) uint64 {
-	tx := transactions.SendTxArgs{
-		From:  (types.Address)(from),
-		Value: (*hexutil.Big)(pathprocessor.ZeroBigIntValue),
-	}
-	switch s {
-	case ENSRegister:
-		estimate, err := ensService.API().RegisterEstimate(context.Background(), network.ChainID, tx, EstimateUsername, EstimatePubKey)
-		if err != nil {
-			return 400000
-		}
-		return estimate
-
-	case ENSRelease:
-		estimate, err := ensService.API().ReleaseEstimate(context.Background(), network.ChainID, tx, EstimateUsername)
-		if err != nil {
-			return 200000
-		}
-		return estimate
-
-	case ENSSetPubKey:
-		estimate, err := ensService.API().SetPubKeyEstimate(context.Background(), network.ChainID, tx, fmt.Sprint(EstimateUsername, ".stateofus.eth"), EstimatePubKey)
-		if err != nil {
-			return 400000
-		}
-		return estimate
-
-	case StickersBuy:
-		packID := &bigint.BigInt{Int: big.NewInt(2)}
-		estimate, err := stickersService.API().BuyEstimate(context.Background(), network.ChainID, (types.Address)(from), packID)
-		if err != nil {
-			return 400000
-		}
-		return estimate
-
-	default:
-		return 0
-	}
 }
